Fixed Resizing

+++ Messaging/ZYMessaging.h
@@ -13,10 +13,10 @@ typedef struct {
 	BOOL shouldForceSize;
 	// Can't use CGSize because it uses CGFloats which aren't able to be transferred between 32/64bit processes (because its float in one and something else (double? i can't remember) in the other).
 	// Also why we can't use CGFloat here
-	float wantedClientOriginX;
-	float wantedClientOriginY;
-	float wantedClientWidth;
-	float wantedClientHeight;
+	CGFloat wantedClientOriginX;
+	CGFloat wantedClientOriginY;
+	CGFloat wantedClientWidth;
+	CGFloat wantedClientHeight;
 
 	BOOL statusBarVisibility;
 	BOOL shouldForceStatusBar;
